+++ src/modules/common.c
@@ -2,12 +2,12 @@
 #include <stdio.h>
 #include "../defs.h"
 
-void print_output(float perc, Options options)
+void print_output(double perc, Options options)
 {
 	const char *fmt;
 
 	if (options.round) {
-		perc = roundf(perc);
+		perc = round(perc);
 		fmt = options.perc == 1 ? "%.0f%%\n" : "%.0f\n";
 	} else {
 		fmt = options.perc == 1 ? "%.2f%%\n" : "%.2f\n";
